+++ src/utils/nodejs.rs
@@ -1,16 +1,15 @@
-use std;
-use std::env::consts::{ARCH, OS};
-use std::path::PathBuf;
-use std::string::String;
+use std::{
+    env::consts::{ARCH, OS},
+    io::{self, Read},
+    path::PathBuf,
+};
 
 use serde_derive::Deserialize;
 use serde_json;
 
-use crate::lib::version;
-use crate::utils;
+use crate::{lib::version, utils};
 
 const DIST_JSON_URL: &str = "https://nodejs.org/dist/index.json";
-const ERROR_MSG: &str = "error: utils: nodejs";
 
 #[derive(Debug, Deserialize)]
 pub struct Release {
@@ -81,11 +80,12 @@ pub fn install_path() -> PathBuf {
     utils::env::home_dir().join(".local").join("node")
 }
 
-pub fn latest_version() -> String {
+pub fn latest_version() -> io::Result<String> {
     let req = utils::http::create_request(DIST_JSON_URL, None);
-    let mut res = utils::http::fetch_request(req).expect(ERROR_MSG);
-    let body = res.text().expect(ERROR_MSG);
-    let releases: Vec<Release> = serde_json::from_str(&body).expect(ERROR_MSG);
+    let mut res = utils::http::fetch_request(req)?;
+    let mut body = String::new();
+    res.read_to_string(&mut body)?;
+    let releases: Vec<Release> = serde_json::from_str(&body)?;
 
     let latest_release: &Release = releases
         .iter()
@@ -102,7 +102,7 @@ pub fn latest_version() -> String {
         })
         .unwrap();
 
-    String::from(latest_release.version.as_str().trim())
+    Ok(String::from(latest_release.version.as_str().trim()))
 }
 
 pub fn lib_dir() -> PathBuf {
@@ -140,7 +140,7 @@ mod tests {
 
     #[test]
     fn latest_version_found() {
-        let version = latest_version();
+        let version = latest_version().expect("must fetch");
         assert!(version.starts_with('v'));
     }